requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
drf-extensions==0.5.0
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
Markdown==3.1.1
mccabe==0.6.1
mypy==0.740
mypy-extensions==0.4.3
pathspec==0.6.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
regex==2019.11.1
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
